working on torrent pi disk space rpt
